Fix insufficient material resulting in UI freeze
Prior to this commit, if material was sufficiently small, there were
cases where creating queries became impossible. However, the code would
keep trying (and failing), which would result in a CPU spike and a
brower UI that would appear to freeze.

This commit will allow for limited reattempts at query creation, then
will fail with an alert to the user recommending they expand the
material.