Fix bug in query generation for Finish
In the Finish query generation algorithm, if a verse was found that
didn't have a valid Finish from it, the algorithm would try again but
would keep using the same verse. That's fixed now.

Also, if "bible" is undefined, the Query class used to let the Material
class select the next bible. This seemed normally correct (and when
using a generated distribution is actually fine); however, if not using
a distribution, on verse selection initial failure in a
multi-translation setup, the bible would move on to the next, which is
not desired. So now "create()" explicitly selects the next bible if it's
not defined/provided.