View past query shows wrong query fixed
When 83bc4fd (or roughly then) was implemented, quotes no longer have a
set bible. This caused the #setup_query function to fail in a curious
way. Only after a quote query, all subsequent queries would not retain
their query during #build_board. The reason is that #setup_query will
try to populate a query from the query_cache if a query does not already
exist. But it also checks the cache for a query of the same type and
bible. When there's no bible, it assumed it should drop the cache and
just create new queries.

This tiny patch causes the correct query to be found in the cache. It
does this by ignoring bible when the record has no bible.